requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
protobuf==3.20.3
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
